chore: Eliminate warnings in the Python pipeline (#182)
* FutureWarning about line_terminator in pandas to_csv

* Replace deprecated op parameter in geopandas sjoin

* Replace deprecated np.int by int

* Replace deprecated np.float by float

* Replace deprecated np.str

* Avoid SettingWithCopyWarning

* Avoid FutureWarning about DataFrameGroupBy.sum

* Resolve FutureWarning about DataFrameGroupBy.sum

* Resolve FutureWarning about inplace parameter in pandas.Categorical.add_categories

* Code fix

* Treat warnings as errors

* Try to increase verbosity

* Trying to raise errors for warnings

* Remove warning

* Avoid GeoDataFrame concatenation warnings

* Fix typo

* Change structure to avoid warnings of creating empty GeoDataFrame

* Replace np.bool to avoid DeprecationWarning

* Starting to catch warnings in 3rd party packages

+++ data/hts/edgt_lyon/cleaned_adisp.py
@@ -42,11 +42,11 @@ def execute(context):
 
     # Transform original IDs to integer (they are hierarchichal)
     df_households["edgt_household_id"] = (df_households["ZFM"] + df_households["ECH"]).astype(int)
-    df_persons["edgt_person_id"] = df_persons["PER"].astype(np.int)
+    df_persons["edgt_person_id"] = df_persons["PER"].astype(int)
     df_persons["edgt_household_id"] = (df_persons["ZFP"] + df_persons["ECH"]).astype(int)
-    df_trips["edgt_person_id"] = df_trips["PER"].astype(np.int)
+    df_trips["edgt_person_id"] = df_trips["PER"].astype(int)
     df_trips["edgt_household_id"] = (df_trips["ZFD"] + df_trips["ECH"]).astype(int)
-    df_trips["edgt_trip_id"] = df_trips["NDEP"].astype(np.int)
+    df_trips["edgt_trip_id"] = df_trips["NDEP"].astype(int)
 
     # Construct new IDs for households, persons and trips (which are unique globally)
     df_households["household_id"] = np.arange(len(df_households))
@@ -67,11 +67,11 @@ def execute(context):
     df_trips = hts.compute_first_last(df_trips)
 
     # Weight
-    df_persons["person_weight"] = df_persons["COEP"].astype(np.float)
-    df_households["household_weight"] = df_households["COE0"].astype(np.float)
+    df_persons["person_weight"] = df_persons["COEP"].astype(float)
+    df_households["household_weight"] = df_households["COE0"].astype(float)
 
     # Clean age
-    df_persons["age"] = df_persons["P4"].astype(np.int)
+    df_persons["age"] = df_persons["P4"].astype(int)
 
     # Clean sex
     df_persons.loc[df_persons["P2"] == 1, "sex"] = "male"
@@ -107,8 +107,8 @@ def execute(context):
 
     # Number of vehicles
     df_households["number_of_vehicles"] = df_households["M6"] + df_households["M14"]
-    df_households["number_of_vehicles"] = df_households["number_of_vehicles"].astype(np.int)
-    df_households["number_of_bikes"] = df_households["M21"].astype(np.int)
+    df_households["number_of_vehicles"] = df_households["number_of_vehicles"].astype(int)
+    df_households["number_of_bikes"] = df_households["M21"].astype(int)
 
     # License
     df_persons["has_license"] = df_persons["P7"] == "1"
